requirements.txt
graphene==2.1.8
graphene-django==2.10.0
python-graphql-client==0.2.0
invoke==1.4.1
django==3.0.6
django-graphql-jwt==0.3.1
django-cors-headers==3.3.0
gql-query-builder==0.1.5
PyYAML==5.3.1
jsonpath-ng==1.5.1



